requirements.txt
# Dependencias del Sistema Financiero SURTHILANAS

# Bot de Telegram
python-telegram-bot==20.7
python-telegram-bot[job-queue]==20.7

# Google Drive y Sheets
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==5.12.3
oauth2client==4.1.3

# Utilidades
python-dotenv==1.0.0
pytz==2023.3

# Manejo de datos
python-dateutil==2.8.2
